backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
install==1.3.5
Markdown==3.4.1
numpy==1.24.1
pandas==1.5.2
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
